Add support for the old, deprecated MCP23016 expander, too.
The main difference with the '17 is that it lacks advanced interrupt
control features, and pull-up configuration.

This includes changes to DigitalInOut to raise ValueError when trying to
set pull-up configuration as well as pull-down, on the MCP23016 only.

Note that for compatibility with the already existing '17 module, and for
an eventual '18 module, the ports are called A and B, rather than 0 and 1.